services/authservice/requirements.txt
-i https://pypi.org/simple
asgiref==3.9.1; python_version >= '3.9'
django==5.2.6; python_version >= '3.10'
django-cors-headers==4.8.0; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
django-prometheus==2.4.1; python_version >= '3.10'